used styled-components for styling
